LPD-31781 Check node-scripts hash in preflight
