+++ modules/_node-scripts/format/self.mjs
@@ -3,12 +3,11 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import crypto from 'crypto';
-import fg from 'fast-glob';
 import fs from 'fs/promises';
 import path from 'path';
-import url from 'url';
 
+import {NODE_SCRIPTS_PATH} from '../util/constants.mjs';
+import digestNodeScripts from '../util/digestNodeScripts.mjs';
 import getNamedArguments from '../util/getNamedArguments.mjs';
 import objectSF from '../util/objectSF.mjs';
 import mainBase from './index.mjs';
@@ -24,13 +23,11 @@ export default async function main() {
 
 	await mainBase();
 
-	const __dirname = path.dirname(url.fileURLToPath(import.meta.url));
-	const nodeScriptsPath = path.resolve(__dirname, '..');
-	const packageJSONPath = path.join(nodeScriptsPath, 'package.json');
+	const packageJSONPath = path.join(NODE_SCRIPTS_PATH, 'package.json');
 
 	const packageJSON = JSON.parse(await fs.readFile(packageJSONPath, 'utf-8'));
 
-	const expectedHash = await digestNodeScripts(nodeScriptsPath);
+	const expectedHash = await digestNodeScripts();
 
 	if (packageJSON['com.liferay']['sha256'] === expectedHash) {
 		return;
@@ -52,31 +49,7 @@ Please update the hash field of the package.json file.
 		packageJSON['com.liferay']['sha256'] = expectedHash;
 
 		await fs.writeFile(packageJSONPath, objectSF(packageJSON), 'utf-8');
+
+		console.log('🔐 Updated sha256 field of package.json');
 	}
-}
-
-async function digestNodeScripts(nodeScriptsPath) {
-	const sha256 = crypto.createHash('sha256');
-
-	let files = await fg(['**/*.mjs', '**/*.js'], {
-		absolute: true,
-		cwd: nodeScriptsPath,
-		ignore: ['bundle/sass/binary/**', 'node_modules/**'],
-	});
-
-	files = files.filter((file) => !path.basename(file).startsWith('.'));
-
-	files.sort();
-
-	const fileContents = await Promise.all(
-		files.map(async (file) => {
-			return await fs.readFile(file, 'utf-8');
-		})
-	);
-
-	for (const fileContent of fileContents) {
-		sha256.update(fileContent);
-	}
-
-	return sha256.digest('hex');
 }
